pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["Izpodvypodverta <izpodvypodwerta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
loguru = "^0.7.2"
asyncpg = "^0.29.0"
fastapi-users = "^13.0.0"
sqlalchemy = "^2.0.35"
fastapi-users-db-sqlalchemy = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
